refactor: clean up footer code by dividing into smaller components 🧹🦶
