+++ package.json
@@ -16,6 +16,7 @@
     "bootstrap": "^5.3.3",
     "jwt-decode": "^4.0.0",
     "jwt-decoder": "^0.0.0",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-loader-spinner": "^6.1.6",
