+++ src/components/Footer/components/FooterMenu/FooterBrand/components/FooterBrandDescription.jsx
@@ -1,0 +1,11 @@
+function FooterBrandDescription() {
+  return (
+    <div className=" col col-xl-6 ">
+      <p className="text-white mx-sm-auto mx-lg-auto">
+        Multipurpose eCommerce website
+      </p>
+    </div>
+  );
+}
+
+export default FooterBrandDescription;
