+++ src/components/Footer/components/FooterMenu/FooterBrand/components/FooterLogo.jsx
@@ -1,0 +1,13 @@
+import whiteLogo from "./../../../assets/images/whiteLogo.svg";
+
+function FooterLogo() {
+  return (
+    <div className="col col-xl-3">
+      <div className="logo mx-sm-auto mx-lg-auto w-100">
+        <img src={whiteLogo} alt="" />
+      </div>
+    </div>
+  );
+}
+
+export default FooterLogo;
