+++ src/components/Footer/components/FooterMenu/FooterBrand/components/TextDivider.jsx
@@ -1,0 +1,10 @@
+function TextDivider() {
+  return (
+    <div className="col col-xl-1">
+      <div className="vr d-xl-block d-none"></div>
+      <div className="smallLine d-sm-block d-xl-none mx-auto "></div>
+    </div>
+  );
+}
+
+export default TextDivider;
