+++ src/components/Footer/data/footerData.jsx
@@ -1,0 +1,51 @@
+import facebookIcon from "../assets/images/icons/facebookIcon.svg";
+import instgramIcon from "../assets/images/icons/instgramIcon.svg";
+import youtubeIcon from "../assets/images/icons/youtubeIcon.svg";
+
+export const footerCategories = [
+  {
+    name: "home",
+    path: "/",
+  },
+  {
+    name: "categories",
+    path: "/categories",
+  },
+  {
+    name: "products",
+    path: "/products",
+  },
+  {
+    name: "contact us",
+    path: "/contact-us",
+  },
+];
+
+export const legalLinks = [
+  {
+    name: "Privacy Policy",
+    path: "",
+  },
+  {
+    name: "Terms of Use",
+    path: "",
+  },
+];
+
+export const socialLinks = [
+  {
+    name: "instgramIcon",
+    icon: instgramIcon,
+    alt: "Instgram",
+  },
+  {
+    name: "facebookIcon",
+    icon: facebookIcon,
+    alt: "Facebook",
+  },
+  {
+    name: "YoutubeIcon",
+    icon: youtubeIcon,
+    alt: "Youtube",
+  },
+];
